[FEATURE] Kickstart FLOW3 Coding Standard
During the Core Team Meeting at TYPO3 Developer Days 2012
there was the idea to kickstart the FLOW3 coding standard.

The included sniffs are a first proposal.
They based on the FLOW3 coding guidelines and useful sniffs
which are not covered by the CGL.

This patch is the first one which some basics sniffs.
The list of sniffs is not final and has to discuss with the
FLOW3 core team.

Change-Id: I2fb2fa327935b0a03ed978a0c2e19d2c49524b79