Create a working PEAR package
The PEAR package was broken and contains only the package.xml file but no further files like the ruleset.xml.

This revision contains a new working package and it reflects the naming change of the standard as well.

This means that the user have to uninstall the old package by
$ sudo pear uninstall typo3/PHPCS_FLOW3_Standard
and install the new one:
$ sudo pear install typo3/TYPO3Flow

PEAR will automatically install the required typo3/TYPO3SniffPool package.